requirements.txt
# packages
requests
kafka-python
pandas
streamlit
langchain-core
langchain-community
llama-index
scikit-learn
mlflow
bentoml
fastapi
pydantic
pytest
delta-spark==2.4.0
dbt-core
apache-airflow
prometheus-client
elasticsearch
jaeger-client==4.8.0
numexpr
numpy
pandas
pydantic
scikit-learn>=1.0.0 ; extra == 'pipelines'
matplotlib
seaborn
sqlalchemy
ipython
pyzmq
panel
xarray
logbook
minimal-snowplow-tracker
flask-appbuilder>=4.3.11
zope-interface
monotonic
msgpack
sortedcontainers
pyasn1-modules
pyasn1
pooch
requests-mock
appdirs
colorcet
PyYAML
joblib
packaging
protobuf
pyarrow
more-itertools
connexion
starlette
anyio
pathspec
sqlparse
urllib3
langchain-groq
alpha_vantage
twelvedata